Add CTE support in IVM
Simple CTEs which does not contain aggregates or DISTINCT are now
supported similarly to simple sub-queries.

Before a view is maintained, all CTEs are converted to corresponding
subqueries to enable to treat CTEs as same as subqueries. For this
end, codes of the static fnction inline_cte in the core
(optimizer/plan/subselect.c) was imported.