Remove a unnecessary function call
It was added by 3de95c09 for consistency with the patch version
proposed for PostgreSQL core, but the called function was not
one we intended. Fortunately, the behavior has not been affected.

Now, I decided to remove this funciton since I come to think it is
confusable.

Also, added some comments on an argument left for the similar
consistency.