Add support for outer joins (#48)
Outer joins can produce dangling tuples, which are tuples from the
preserved (non-nullable) side that are null-extended in the attributes
of the non-preserved (nullable) side when no matching rows exist.
In incremental maintenance of outer join views, it is necessary to
handle changes to dangling tuples. Such tuples need to be inserted
when a base-table tuple is deleted, and deleted when a base-table
tuple is inserted.

To analyze how dangling tuples are affected by a table modification,
an outer join query is transformed into a normalized form that
consists of one or more terms. The normalized form is a bag union of
terms, and each term is an inner join of some base tables, which is
usually null-extended due to one or more anti-joins with other base
tables.

For each term in the normalized form, if the modified table is among
the inner-joined tables, tuples belonging to that term should be
inserted into the view when tuples are inserted into the table, or
deleted when tuples are deleted. Such changes are called the primary
delta, and can be computed similarly to those for normal inner-join
views after rewriting the query to reduce the strength of any outer
join that includes the modified table.

Otherwise, if the modified table is among the anti-joined tables of a
term, tuples belonging to that term may be deleted from the view on
insertions, or inserted on deletions, depending on whether tuples that
satisfy the join qual between the inner-join and anti-join parts exist.
Such changes are called the secondary delta, and are computed and
applied per term that needs to be modified. The precomputed primary
delta can be used to generate dangling tuples to be inserted, or to
build the condition for deleting dangling tuples for each term.

Restrictions:

- The target list of an outer join view must contain attributes used in join conditions.
- The target list of an outer join view cannot contain non-strict functions.
- Only simple equijoins are supported.
- The WHERE clause of an outer join view cannot contain non-null-rejecting predicates.
- Aggregates are not supported with outer joins.
- Subqueries are not supported with outer joinx.